Fixed test_model_output of keras implementation
Changed the format of intermidiate representaiton of features from dict
to list (Invoking keras layer only accepts Tensor/tuple/list)

Converted the code which adds resample feature maps to backbone features
to the keras layer ResampleFeatureAdder (needed to compile keras model)

Converted function fuse_features to keras layer FuseFeature

Added methods which test that the layers of the keras model are compilable.

Changed representation of model in test_model_output function from
keras model to chain of keras layers because the output of the keras model is
different to the output of the previous model. (can this be the cause of
keras model not using session graph?, so its not using the set random seed?)