Fixed name compatibility according to restoring model from checkpoint
Changed dict keys from int to str according to this comment
https://github.com/tensorflow/tensorflow/issues/36916#issuecomment-640395999

Added setting names of conv and bn backbone's layers.
Varible names of the keras model different depends on having context of
tf.Graph().as_default.
When I tried to restore the model from the
checkpoint it failed cause graph contains two nodes with same name
efficientnet-b0/blocks_0/se/conv2d/bias:0 but with different shape (32,)
and (8,). In the same time name compatibility test passed. I found that
the difference between the test and my useage of keras.Model was that in
test we created the keras.Model in tf.Graph().as_default context and I
used it without this context. After removing the context I found that
the name of blocks incorrect